CHG: refactiring cata2data
