+++ pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["setuptools>=65.5.0", "setuptools-scm", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "cata2data"
+authors = [
+    {name = "Micah Bowles", email = "micah.bowles@postgrad.manchester.ac.uk"},
+]
+maintainers = [
+    {name = "Micah Bowles", email = "micah.bowles@postgrad.manchester.ac.uk"},
+]
+description = "Produce a loadable data set from a catalogue"
+readme = "README.md"
+license = {text = "GNU GPL"}
+requires-python = ">=3.9"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Operating System :: OS Independent"
+]
+dynamic = ["version", "dependencies"]
+
+[project.urls]
+Source = "https://github.com/mbowles/cata2data"
+
+[tool.setuptools]
+include-package-data = false
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["cata2data*"]
+namespaces = false
+
+[tool.setuptools_scm]
+write_to = "src/_version.py"
