+++ requirements.txt
@@ -1,0 +1,4 @@
+numpy==1.24.2
+pandas==1.5.3
+matplotlib==3.6.3
+astropy==5.2.1
