+++ setup.cfg
@@ -1,0 +1,18 @@
+[metadata]
+name = pyralysis
+version = attr: pyralysis.__version__
+description = PYthon Radio Astronomy anaLYSis and Image Synthesis
+long_description = file: README.md
+long_description_content_type = text/markdown
+
+[yapf]
+based_on_style = pep8
+column_limit = 100
+dedent_closing_brackets = true
+allow_split_before_dict_value = false
+split_before_bitwise_operator = false
+
+[isort]
+multi_line_output = 3
+include_trailing_comma = true
+line_length = 100
