+++ src/cata2data/__init__.py
@@ -1,0 +1,1 @@
+from .cata2data import *
